requirements.txt
google-cloud-aiplatform
python-dotenv